+++ Runtime/UnityEditor/MonoBehaviours/ActiveSystemsViewer.cs
@@ -1,3 +1,5 @@
+using EcsRx.Collections;
+using EcsRx.Collections.Database;
 using EcsRx.Executor;
 using UnityEngine;
 using Zenject;
@@ -8,5 +10,11 @@ namespace EcsRx.UnityEditor.MonoBehaviours
     {
         [Inject]
         public ISystemExecutor SystemExecutor { get; private set; }
+
+        [Inject]
+        public IEntityDatabase EntityDatabase { get; private set; }
+
+        [Inject]
+        public IObservableGroupManager ObservableGroupManager { get; private set; }
     }
 }