+++ Samples~/BCKWorksFramework/Engine/Defines/ReactiveProperties/SceneStateReactiveProperty.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using UniRx;
+
+namespace BCKWorks.Engine.Defines
+{
+    [Serializable]
+    public class SceneStateReactiveProperty : ReactiveProperty<SceneStateType>
+    {
+        public SceneStateReactiveProperty()
+            : base()
+        {
+
+        }
+
+        public SceneStateReactiveProperty(SceneStateType initialValue)
+            : base(initialValue)
+        {
+
+        }
+    }
+}