+++ Samples~/BCKWorksFramework/Engine/Events/LoadingFinishedEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace BCKWorks.Engine.Events
+{
+    public class LoadingFinishedEvent
+    {
+        public string TargetSceneName;
+    }
+}