+++ Samples~/BCKWorksFramework/Engine/Events/LoadingPreparingEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace BCKWorks.Engine.Events
+{
+    public class LoadingPreparingEvent
+    {
+        public string ChangeToSceneName;
+    }
+}