+++ Samples~/BCKWorksFramework/Engine/Modules/Status/ADT/IEngineStatus.cs
@@ -1,0 +1,10 @@
+﻿using BCKWorks.Engine.Defines;
+
+namespace BCKWorks.Engine.Modules.Status
+{
+    public interface IEngineStatus
+    {
+        EngineStateType State { get; set; }
+        EngineStateReactiveProperty StateRP { get; }
+    }
+}