+++ Samples~/BCKWorksFramework/Engine/Modules/Status/ADT/ILoadingStatus.cs
@@ -1,0 +1,10 @@
+﻿using BCKWorks.Engine.Defines;
+
+namespace BCKWorks.Engine.Modules.Status
+{
+    public interface ILoadingStatus
+    {
+        LoadingStateType State { get; set; }
+        LoadingStateReactiveProperty StateRP { get; }
+    }
+}