+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 All notable changes to this package will be documented in this file.
 
-## [2.1.2] - 2020-06-09
+## [2.1.2] - 2020-06-20
 ### Added 
-- Examples
+- ConvertEntity in GameObjectExtensions.cs
+- ComponentUIAspect added default string label when not defined properties up
+- ComponentUIAspect added LayerMaskEditorInput
 
 ## [2.1.1] - 2020-06-08
 ### Added